+++ src/components/Button/Button.svelte
@@ -17,22 +17,24 @@
   export let flat = false;
   export let iconClass = "";
   export let color = "primary";
+  export let href = null;
+  export let fab = false;
 
   export let remove = "";
   export let add = "";
   export let replace = {};
 
-  let commonDefault = 'py-2 px-4 uppercase text-sm font-medium relative overflow-hidden';
-  let basicDefault = 'text-white transition';
-  let outlinedDefault = 'bg-transparent border border-solid';
-  let textDefault = 'bg-transparent border-none px-3 hover:bg-transparent';
-  let iconDefault = 'p-4 m-4 flex items-center';
-  let fabDefault = 'px-4 hover:bg-transparent';
-  let smallDefault = 'p-1 h-4 w-4';
-  let disabledDefault = 'bg-gray-300 text-gray-500 elevation-none pointer-events-none hover:bg-gray-300 cursor-default';
-  let elevationDefault = 'hover:elevation-5 elevation-3';
+  const classesDefault = 'py-2 px-4 uppercase text-sm font-medium relative overflow-hidden';
+  const basicDefault = 'text-white transition';
+  const outlinedDefault = 'bg-transparent border border-solid';
+  const textDefault = 'bg-transparent border-none px-4 hover:bg-transparent';
+  const iconDefault = 'p-4 flex items-center';
+  const fabDefault = 'hover:bg-transparent';
+  const smallDefault = 'p-1 h-4 w-4';
+  const disabledDefault = 'bg-gray-300 text-gray-500 dark:bg-dark-400 elevation-none pointer-events-none hover:bg-gray-300 cursor-default';
+  const elevationDefault = 'hover:elevation-5 elevation-3';
 
-  export let commonClasses = commonDefault;
+  export let classes = classesDefault;
   export let basicClasses = basicDefault;
   export let outlinedClasses = outlinedDefault;
   export let textClasses = textDefault;
@@ -42,12 +44,12 @@
   export let disabledClasses = disabledDefault;
   export let elevationClasses = elevationDefault;
 
-  const fab = text && icon;
+  fab = fab || (text && icon);
   const basic = !outlined && !text && !fab;
   const elevation = (basic || icon) && !disabled && !flat && !text;
   
-  let classes = "";
-  let iClasses = "";
+  let Classes = i => i;
+  let iClasses = i => i;
   let shade = 0;
 
   $: {
@@ -63,42 +65,45 @@
     txt,
   } = utils(color);
 
-  const cb = new ClassBuilder(commonClasses, commonDefault);
+  const cb = new ClassBuilder(classes, classesDefault);
   let iconCb;
 
   if (icon) {
     iconCb = new ClassBuilder(iconClass);
   }
 
-  $: {
-      classes = cb
-        .flush()
-        .add(basicClasses, basic, basicDefault)
-        .add(`${bg(normal)} hover:${bg(lighter)}`, basic)
-        .add(elevationClasses, elevation, elevationDefault)
-        .add(outlinedClasses, outlined, outlinedDefault)
-        .add(
-          `${border(lighter)} ${txt(normal)} hover:${bg("trans")}`,
-          outlined)
-        .add(`${txt(lighter)}`, text)
-        .add(textClasses, text, textDefault)
-        .add(iconClasses, icon, iconDefault)
-        .remove('py-2', icon)
-        .add(fabClasses, fab, fabDefault)
-        .remove(txt(lighter), fab)
-        .add(disabledClasses, disabled, disabledDefault)
-        .add(smallClasses, small, smallDefault)
-        .add('flex items-center justify-center', small && icon)
-        .remove(remove)
-        .replace(replace)
-        .add(add)
-        .get();
-  }
+  $: classes = cb
+      .flush()
+      .add(basicClasses, basic, basicDefault)
+      .add(`${bg(normal)} hover:${bg(lighter)}`, basic)
+      .add(elevationClasses, elevation, elevationDefault)
+      .add(outlinedClasses, outlined, outlinedDefault)
+      .add(
+        `${border(lighter)} ${txt(normal)} hover:${bg("trans")} dark-hover:${bg("transDark")}`,
+        outlined)
+      .add(`${txt(lighter)}`, text)
+      .add(textClasses, text, textDefault)
+      .add(iconClasses, icon, iconDefault)
+      .remove("py-2", icon)
+      .remove(txt(lighter), fab)
+      .add(disabledClasses, disabled, disabledDefault)
+      .add(smallClasses, small, smallDefault)
+      .add("flex items-center justify-center", small && icon)
+      .add("border-solid", outlined)
+      .add("rounded-full", icon)
+      .add("w-full", block)
+      .add("rounded", basic || outlined || text)
+      .add("button", !icon)
+      .add(fabClasses, fab, fabDefault)
+      .add(`hover:${bg("transLight")}`, fab)
+      .add(className)
+      .remove(remove)
+      .replace(replace)
+      .add(add)
+      .get();
 
-  $: {
-    if (iconCb) {
-      iClasses = iconCb.flush().add(txt(), fab && !iconClass).get();
-    }
+  $: if (iconCb) {
+    iClasses = iconCb.flush().add(txt(), fab && !iconClass).get();
   }
 
   const ripple = createRipple((text || fab || outlined) ? color : "white");
@@ -114,23 +119,48 @@
     'light',
     'dark',
     'flat',
+    'add',
+    'remove',
+    'replace',
   ], $$props);
 </script>
 
-<button
-  use:ripple
-  class:border-solid={outlined}
-  class:rounded-full={icon}
-  class:w-full={block}
-  class:rounded={basic || outlined || text}
-  class="{classes} {className} button"
-  class:button={!icon}
-  on:click
-  {...props}
-  {disabled}
-  on:click={() => (value = !value)}>
-  {#if icon}
-    <Icon class={iClasses} {small}>{icon}</Icon>
-  {/if}
-  <slot />
-</button>
+
+{#if href}
+  <a
+    {href}
+    {...props}
+  >
+    <button
+      use:ripple
+      class={classes}
+      {...props}
+      {disabled}
+      on:click={() => (value = !value)}
+      on:click
+      on:mouseover
+      on:*
+    >
+      {#if icon}
+        <Icon class={iClasses} {small}>{icon}</Icon>
+      {/if}
+      <slot />
+    </button>
+  </a>
+{:else}
+  <button
+    use:ripple
+    class={classes}
+    {...props}
+    {disabled}
+    on:click={() => (value = !value)}
+    on:click
+    on:mouseover
+    on:*
+  >
+    {#if icon}
+      <Icon class={iClasses} {small}>{icon}</Icon>
+    {/if}
+    <slot />
+  </button>
+{/if}