+++ src/components/Card/Card.svelte
@@ -1,12 +1,28 @@
 <script>
+  import { ClassBuilder } from "../../utils/classes.js";
+
+  const classesDefault = "rounded inline-flex flex-col overflow-hidden transition";
+
   export let hover = true;
-  let className = "";
+  export let elevation = 1;
+  export let hoverElevation = 8;
+  export let classes = classesDefault;
+
   export {className as class};
-  export let wrapperClasses = "rounded inline-flex flex-col overflow-hidden";
+  let className = "";
+
+  const cb = new ClassBuilder(classes, classesDefault);
+
+  $: c = cb
+    .flush()
+    .add(`elevation-${elevation} hover:elevation-${hoverElevation}`, hover)
+    .add(classes, true, classesDefault)
+    .add(className)
+    .get();
 </script>
 
 <div
-  class="{className} {hover ? 'elevation-1 hover:elevation-8' : ''} {wrapperClasses}">
+  class={c}>
   <slot name="title" />
   <slot name="media" />
   <slot />