+++ src/components/List/List.svelte
@@ -1,52 +1,76 @@
 <script>
+  import { ClassBuilder } from "../../utils/classes.js";
+ 
   import ListItem from "./ListItem.svelte";
 
   export let items = [];
-  export let item = {};
   export let value = "";
-  export let text = "";
   export let dense = false;
-  export let navigation = false;
   export let select = false;
+
+  export const level = null;
+  export const text = "";
+  export const item = {};
+  export const to = null;
+  export const selectedClasses = i => i;
+  export const itemClasses = i => i;
+
+  const classesDefault = "py-2 rounded";
+
+  export let classes = classesDefault;
+
   let className = "";
   export {className as class};
 
-  export let listClasses = "py-2 rounded";
+  function id(i) {
+    if (i.id !== undefined) return i.id;
+    if (i.value !== undefined) return i.value;
+    if (i.to !== undefined) return i.to;
+    if (i.text !== undefined) return i.text;
+    return i;
+  }
 
-  const props = {
-    dense,
-    navigation
-  };
+  function getText(i) {
+    if (i.text !== undefined) return i.text;
+    if (i.value !== undefined) return i.value;
+    return i;
+  }
 
-  const id = item => item.id || item.value || item.to || item.text || item;
+  const cb = new ClassBuilder(className);
+
+  $: c = cb
+    .flush()
+    .add(classes, true, classesDefault)
+    .add(className)
+    .get();
 </script>
 
-<div class={className}>
-  <ul class={listClasses} class:rounded-t-none={select}>
-    {#each items as item, i}
-      {#if item.to}
-        <slot name="item" {item} {dense} {navigation} {value}>
-          <a tabindex={i + 1} href={item.to}>
-            <ListItem bind:value {...item} id={id(item)} {...props} on:change>
-              {item.text}
-            </ListItem>
-          </a>
-        </slot>
-      {:else}
-        <slot name="item" {item} {dense} {navigation} {value}>
-          <ListItem
-            bind:value
-            {...item}
-            tabindex={i + 1}
-            id={id(item)}
-            selected={value === id(item)}
-            {...props}
-            on:change
-            on:click>
-            {item.text || item.value || item}
+<ul class={c} class:rounded-t-none={select}>
+  {#each items as item, i}
+    {#if item.to !== undefined}
+      <slot name="item" {item} {dense} {value}>
+        <a tabindex={i + 1} href={item.to}>
+          <ListItem bind:value {...item} id={id(item)} {dense} on:change>
+            {item.text}
           </ListItem>
-        </slot>
-      {/if}
-    {/each}
-  </ul>
-</div>
+        </a>
+      </slot>
+    {:else}
+      <slot name="item" {item} {dense} {value}>
+        <ListItem
+          bind:value
+          {selectedClasses}
+          {itemClasses}
+          {...item}
+          tabindex={i + 1}
+          id={id(item)}
+          selected={value === id(item)}
+          {dense}
+          on:change
+          on:click>
+          {getText(item)}
+        </ListItem>
+      </slot>
+    {/if}
+  {/each}
+</ul>