+++ src/components/RadioButton/RadioButton.svelte
@@ -1,6 +1,12 @@
 <script>
   import Icon from "../Icon";
   import Ripple from "../Ripple";
+  import Label from "../Checkbox/Label.svelte";
+  import { ClassBuilder } from "../../utils/classes.js";
+
+  const classesDefault = "inline-flex block items-center mb-2 cursor-pointer z-0";
+  let className = "";
+  export {className as class};
 
   export let selected = "";
   export let label = "";
@@ -8,7 +14,8 @@
   export let disabled = false;
   export let name = "";
   export let value = "";
-  export let wrapperClasses = "inline-flex block items-center mb-2 cursor-pointer z-0";
+  export let classes = classesDefault;
+  export let labelClasses = i => i;
 
   function select() {
     if (disabled) return;
@@ -16,17 +23,25 @@
     selected = value;
   }
 
+  const cb = new ClassBuilder(classes, classesDefault);
+  $: c = cb
+    .flush()
+    .add(classes, true, classesDefault)
+    .add(className)
+    .get();
+
   $: rippleColor = value && !disabled ? color : 'gray';
 </script>
 
 <div
-  class={wrapperClasses}
+  class={c}
   on:click={select}>
   <input
     aria-label={label}
     class="hidden"
     type="radio"
     role="radio"
+    {name}
     selected={selected === value} />
   <div class="relative">
     <Ripple color={rippleColor}>
@@ -35,17 +50,13 @@
           radio_button_checked
         </Icon>
       {:else}
-        <Icon class={disabled ? 'text-gray-500' : 'text-gray-600'}>
+        <Icon class={disabled ? 'text-gray-500 dark:text-gray-600' : 'text-gray-600'}>
           radio_button_unchecked
         </Icon>
       {/if}
     </Ripple>
   </div>
-  <label
-    aria-hidden="true"
-    class="pl-2"
-    class:text-gray-500={disabled}
-    class:text-gray-700={!disabled}>
-    {label}
-  </label>
+  <slot name="label">
+    <Label {disabled} {label} class={labelClasses} />
+  </slot>
 </div>