+++ src/docs/PropsTable.svelte
@@ -1,6 +1,8 @@
 <script>
   import DataTable from "../components/DataTable";
 
+  let className = "my-10 w-full";
+  export { className as class };
   export let data = [];
   export let columns = [
     { field: "prop" },
@@ -11,6 +13,7 @@
 </script>
 
 <DataTable
+  class={className}
   editable={false}
   pagination={false}
   sortable={false}