+++ src/routes/components/images.svelte
@@ -1,11 +1,16 @@
 <script>
-  import { Image } from "smelte";
+  import Image from "components/Image";
   import Code from "docs/Code.svelte";
 
   import images from "examples/images.txt";
 
   const range = [...new Array(50)];
 </script>
+
+<p>
+  Smelte includes convenience image component which is useful for lazyloading, but generally we recommend
+  using <a class="a" href="https://github.com/matyunya/svelte-image">Svelte Image</a>.
+</p>
 
 <Code code={images} />